% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackboard_import.R
\name{blackboard_import}
\alias{blackboard_import}
\title{blackboard_import}
\source{
\url{https://help.blackboard.com/Learn/Instructor/Tests_Pools_Surveys/Reuse_Questions/Upload_Questions}
}
\usage{
blackboard_import(fn, out = "out.txt")
}
\arguments{
\item{fn}{The filename of the input file}

\item{out}{The filename of the output file}
}
\value{
NULL. Call this function only for the side effect of creating the output file
}
\description{
Import an exam text file and output a tab delimited version suitable for import into blackboard
}
\details{
Make sure input file is UTF-8 with lines separated by linefeeds (see blackboard docs).
}
\examples{
Questions must be in this format (with one blank line between questions):

1. About 65 million years ago, animals that later evolved into mammals were probably:
a. nocturnal.
b. warm-blooded.
c. specialized chewers.
d. able to turn easily while locomoting on land.
*e. all of the above

\dontrun{
if(interactive()){
 blackboard_import('my exam.txt', 'out.txt')
 }
}
}
