% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_exams.R
\name{create_key}
\alias{create_key}
\title{Create key}
\usage{
create_key(fn)
}
\arguments{
\item{fn}{File name of test with answers marked with *}
}
\value{
A character vector of answers (e.g., DBBACDA...)
}
\description{
Create key vector from test key
}
\details{
Reads an Rmd exam with answers marked with *, and creates a key vector
}
