% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_exams.R
\name{display_scantron}
\alias{display_scantron}
\title{display_scantron}
\usage{
display_scantron(scanfilename, imagedir = "exams")
}
\arguments{
\item{scanfilename}{Filename of scanned image (png format)}

\item{imagedir}{Directory of scanned exam images, Default: 'exams'}
}
\value{
None. Used for side effect of image display.
}
\description{
Utility function to display the scanned exam (in png format) in the Rstudio viewer.
}
\details{
Displays scanned image of exam.
}
\examples{
\dontrun{
if(interactive()){
 display_scantron("exam03.png")
 }
}
}
