% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_test_report.R
\name{email_report}
\alias{email_report}
\title{email_report}
\usage{
email_report(
  scores,
  key,
  roster,
  subject,
  from = "edhagen@wsu.edu",
  imagedir = "exams",
  include_exam = T,
  dry_run = T
)
}
\arguments{
\item{scores}{A data frame with student ID's and exam scores.}

\item{key}{A character vector of correct answers.}

\item{roster}{The roster data frame.}

\item{subject}{A string with the email subject line.}

\item{from}{"from" email address, Default: 'edhagen@wsu.edu'}

\item{imagedir}{A string specifying the directory with the exam images.}

\item{include_exam}{Include the student's exam. Logical.}

\item{dry_run}{Print the body of the emails, but don't actually send them.}
}
\value{
None. Called for side effect of emailing reports.
}
\description{
Emails a report of exam results to each student
}
\details{
Run this function last, after fixing all problems and checking the key.
}
\examples{
\dontrun{
if(interactive()){
 email_report(captured, key)
 }
}
}
\seealso{
\code{\link[glue]{glue_data}}
 \code{\link[mailR]{send.mail}}
 \code{\link[stringr]{str_split}}
}
