% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_test_report.R
\name{email_scores}
\alias{email_scores}
\title{email_scores}
\usage{
email_scores(scores, subject, from = "edhagen@wsu.edu", dry_run = T)
}
\arguments{
\item{scores}{a data frame with a 'Scores' column of scores}

\item{subject}{Email subject line}

\item{from}{Email 'From' address, Default: 'edhagen@wsu.edu'}

\item{dry_run}{Print out email texts to console, Default: T}
}
\value{
Send status of each email
}
\description{
An alternative to email_report to email scores to students
}
\details{
Will email each student in the scores data frame their score
}
\examples{
\dontrun{
if(interactive()){
 #email_scores(d, dry_run = F)
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[stringr]{str_split}}
 \code{\link[emayili]{server}}, \code{\link[emayili]{envelope}}, \code{\link[emayili]{addresses}}, \code{\link[emayili]{subject}}, \code{\link[emayili]{text}}
}
