% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_exams.R
\name{fix_problems}
\alias{fix_problems}
\title{fix_problems}
\usage{
fix_problems(
  captured,
  key,
  roster = roster_to_df(),
  imagedir = "exams",
  out = "captured.corrected.tsv"
)
}
\arguments{
\item{captured}{A data frame of all student answers, as produced by, e.g., import_exams.}

\item{key}{A character vectors of correct answers, as produced by, e.g., import_exams.}

\item{roster}{The student roster in a data frame, Default: roster_to_df()}

\item{out}{Filename for corrected tsv file, Default: 'captured.corrected.tsv'}

\item{number_of_problems}{Number of exam questions.}
}
\value{
A data frame with the corrected student ID's and answers.
}
\description{
Manually fix any problems with unknown student ID numbers,
  possibly blank answers, or multiple marks.
}
\details{
Scanned exams often have artifacts, and students frequently enter an incorrect ID number.
  This function identifies student IDs that do not appear in the roster, double-checks that blank
  answers are actually blank, and allows the instructor to choose one among multiple marks.
}
\examples{
\dontrun{
if(interactive()){
 fix_problems(captured, number_of_problems=25)
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[readr]{write_tsv}}
}
