% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_test_report.R
\name{grades}
\alias{grades}
\title{grades}
\usage{
grades(scores, cutoffs = NULL, probs = c(0.05, 0.1, 0.55, 0.85))
}
\arguments{
\item{scores}{The exam score}

\item{cutoffs}{A named character vector. Names must be "A", "B", "C", "D". Values must
be the minimum score to get that grade. If left NULL, cutoffs will be computed
using the probabilities (probs), Default: NULL}

\item{probs}{The probabilities to compute cutoffs using the quantile function, Default: c(0.05, 0.1, 0.55, 0.85)}
}
\value{
A character vector of simple grades: A, B, C, or D. Plus/minus grades are not available.
}
\description{
Utility function to assign grade based on cutoffs. If
  cutoffs are not supplied, they will be computed from probabilities
}
\details{
This is a utility function that is primarily used to generate midterm grades.
}
\examples{
\dontrun{
if(interactive()){
 grades(captured$SCORE)
 }
}
}
