% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_exams.R
\name{import_exams}
\alias{import_exams}
\title{import_exams}
\usage{
import_exams(
  scan_directory = "exams",
  number_of_questions,
  threshold = c(0.1, 0.5)
)
}
\arguments{
\item{scan_directory}{The directory with the scanned exams in png format, Default: 'exams'}

\item{number_of_questions}{Number of questions on the exam.}
}
\value{
A named list. "key" is a chacter vector of correct answers; "captured" is a
  data frame of student answers.
}
\description{
Import scanned exams (as png files) and produce a key (character vector)
 and student answers (data frame).
}
\details{
A special "bubble sheet" from the exams package should be used. Scan these and
  generate a separate png file for each one. This function will read every file in the
  specified directory and generate a named list with "key", a character vector of correct answers
  corresonding to student ID = "000000000", and "captured", the student answers in a data frame.
}
\examples{
\dontrun{
if(interactive()){
 import_exams(number_of_questions=25)
 }
}
}
\seealso{
\code{\link[exams]{nops_scan}}
 \code{\link[parallel]{detectCores}}
 \code{\link[purrr]{imap_chr}}
 \code{\link[stringr]{str_split}},\code{\link[stringr]{str_c}}
}
