% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_exams.R
\name{import_formreturn}
\alias{import_formreturn}
\title{import_formreturn}
\usage{
import_formreturn(
  exported_data = "captured_data_export.csv",
  number_of_questions
)
}
\arguments{
\item{exported_data}{The filename of the FormReturn csv file, Default: 'captured_data_export.csv'}

\item{number_of_questions}{Number of exam questions.}
}
\value{
A named list. "key" is the answer key as a character vector. "captured" is
 a data frame with STUDENT_ID's and responses to each question.
}
\description{
Import the csv file output by FormReturn into a data frame.
}
\details{
FormReturn generates a csv file with the scanned student IDs and exam responses.
  This function imports that file into a data frame. It identifies the key, which should
  have STUDENT_ID = "00000000", and exports it in a named list as a character vector named "key".
  The student responses are in a data frame named "captured".
}
\examples{
\dontrun{
if(interactive()){
 import_formreturn(number_of_questions=25)
 }
}
}
\seealso{
\code{\link[dplyr]{rename}},\code{\link[dplyr]{filter}}
 \code{\link[readr]{read_tsv}}
}
