% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item analysis.R
\name{itan}
\alias{itan}
\title{itan}
\usage{
itan(x, keys = NULL, digits = 2, no.resp = 5)
}
\arguments{
\item{x}{Student answers (be sure to omit ID, IMAGE, and other extraneous columns)}

\item{keys}{Answer key, Default: NULL}

\item{digits}{Digits to report, Default: 2}

\item{no.resp}{Number of possible responses (e.g., A,B,C,D,E), Default: 5}
}
\value{
Stats for each test item:
\itemize{
   \item Proportion correct
   \item Biserial correlation with total score
   \item Mean score of students who got item correct
   \item Mean score of students who got item incorrect
   \item Number correct
   \item Number omitted
   \item Item response frequencies
}
}
\description{
Item analysis
}
\details{
This function provides some summary statistics for each item on the test.
}
\examples{
\dontrun{
if(interactive()){
 itan(df$captured, df$key)
 }
}
}
