% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_test_report.R
\name{matrix_to_ascii}
\alias{matrix_to_ascii}
\title{matrix_to_ascii}
\usage{
matrix_to_ascii(x)
}
\arguments{
\item{x}{The matrix with columns: Question #, Answer, Key}
}
\value{
A string representation of the matrix.
}
\description{
A utility function to produce an ascii representation of a matrix for inclusion in
  the report emailed to students.
}
\details{
The email report is text-only. This function is used to include a table of incorrect responses.
}
\examples{
\dontrun{
if(interactive()){
 matrix_to_ascii(mat) # matrix with columns: Question #, Answer, Key
 }
}
}
