% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_test_report.R
\name{midterm_grades}
\alias{midterm_grades}
\title{midterm_grades}
\usage{
midterm_grades(captured, cutoffs = NULL)
}
\arguments{
\item{captured}{A data frame with columns: STUDENT_ID, SCORE}

\item{cutoffs}{A named character vector. Names must be "A", "B", "C", "D". Values must
be the minimum score to get that grade. If left NULL, cutoffs will be computed
using the probabilities (probs) in the grades function.}
}
\value{
A csv file with student ids and grades is written to the working directory.
}
\description{
Create a csv file with student ids and grades for WSU midterm grades.
}
\details{
This function creates a file in the correct format for the WSU midterm grade system.
}
\examples{
\dontrun{
if(interactive()){
 midterm_grades(captures)
 }
}
}
\seealso{
\code{\link[readr]{write_csv}}
 \code{\link[stringr]{str_pad}}
}
