% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_test_report.R
\name{plot_scores}
\alias{plot_scores}
\title{plot_scores}
\usage{
plot_scores(e, score)
}
\arguments{
\item{e}{A data frame with all test scores (Score) and all grades (Grade).}

\item{score}{The student's individual score}
}
\value{
A ggplot object
}
\description{
A ggplot of test score distributions, with a vertical line for one student's score.
}
