% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_exams.R
\name{qnames}
\alias{qnames}
\title{qnames}
\usage{
qnames(n)
}
\arguments{
\item{n}{Number of questions}
}
\value{
A character vector of question variable names, from q01...qn
}
\description{
Generate question variable names (e.g., q01, q02, ...)
}
\details{
An internal utility function.
}
\examples{
\dontrun{
if(interactive()){
 qnames(25)
 }
}
}
