% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_exams.R
\name{roster_to_df}
\alias{roster_to_df}
\title{roster_to_df}
\usage{
roster_to_df(filename = "roster.xlsx", out = "roster.tsv")
}
\arguments{
\item{filename}{The roster file name, Default: 'roster.xlsx'}

\item{out}{The output file name of the converted tsv file, Default: 'roster.tsv'}
}
\value{
A data frame with student ids, names, and email addresses.
}
\description{
Imports the WSU portal roster file.
}
\details{
The WSU portal roster file is downloaded in an HTML table but
  with an xls or xlsx file extension (Excel).
}
\examples{
\dontrun{
if(interactive()){
 roster_to_df()
 }
}
}
