% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_exams.R
\name{score_exam}
\alias{score_exam}
\title{score_exam}
\usage{
score_exam(captured, key, cols)
}
\arguments{
\item{captured}{A data frame of answers}

\item{key}{A one-row data frame with the key.}

\item{cols}{The column names of the questions}
}
\value{
A numeric vector of number correct
}
\description{
compute scores by comparing answers to key
}
\details{
A utility function, mostly for internal use, that computes number of correct based on the key.
}
